Split PagesHash from JoinHash
