+++ presto-main/src/main/java/com/facebook/presto/operator/JoinHash.java
@@ -16,166 +16,85 @@ package com.facebook.presto.operator;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PageBuilder;
 import com.google.common.primitives.Ints;
-import io.airlift.units.DataSize;
-import it.unimi.dsi.fastutil.HashCommon;
 import it.unimi.dsi.fastutil.longs.LongArrayList;
 
 import javax.annotation.Nullable;
 
-import java.util.Arrays;
 import java.util.Optional;
 
-import static com.facebook.presto.operator.SyntheticAddress.decodePosition;
-import static com.facebook.presto.operator.SyntheticAddress.decodeSliceIndex;
-import static io.airlift.slice.SizeOf.sizeOf;
-import static io.airlift.units.DataSize.Unit.KILOBYTE;
 import static java.util.Objects.requireNonNull;
 
 // This implementation assumes arrays used in the hash are always a power of 2
 public final class JoinHash
         implements LookupSource
 {
-    private static final DataSize CACHE_SIZE = new DataSize(128, KILOBYTE);
-    private final LongArrayList addresses;
-    private final PagesHashStrategy pagesHashStrategy;
+    private final PagesHash pagesHash;
 
     // we unwrap Optional<JoinFilterFunction> to actual verifier or null in constructor for performance reasons
     // we do quick check for `filterFunction == null` in `getNextJoinPositionFrom` to avoid calls to applyFilterFunction
     @Nullable
     private final JoinFilterFunction filterFunction;
 
-    private final int channelCount;
-    private final int mask;
-    private final int[] key;
-    private final int[] positionLinks;
-    private final long size;
-
-    // Native array of hashes for faster collisions resolution compared
-    // to accessing values in blocks. We use bytes to reduce memory foot print
-    // and there is no performance gain from storing full hashes
-    private final byte[] positionToHashes;
-
     public JoinHash(LongArrayList addresses, PagesHashStrategy pagesHashStrategy, Optional<JoinFilterFunction> filterFunction)
     {
-        this.addresses = requireNonNull(addresses, "addresses is null");
-        this.pagesHashStrategy = requireNonNull(pagesHashStrategy, "pagesHashStrategy is null");
+        this(new PagesHash(addresses, pagesHashStrategy), filterFunction);
+    }
+
+    public JoinHash(PagesHash pagesHash, Optional<JoinFilterFunction> filterFunction)
+    {
+        this.pagesHash = requireNonNull(pagesHash, "pagesHash is null");
         this.filterFunction = requireNonNull(filterFunction, "filterFunction can not be null").orElse(null);
-        this.channelCount = pagesHashStrategy.getChannelCount();
-
-        // reserve memory for the arrays
-        int hashSize = HashCommon.arraySize(addresses.size(), 0.75f);
-
-        mask = hashSize - 1;
-        key = new int[hashSize];
-        Arrays.fill(key, -1);
-
-        this.positionLinks = new int[addresses.size()];
-        Arrays.fill(positionLinks, -1);
-
-        positionToHashes = new byte[addresses.size()];
-
-        // We will process addresses in batches, to save memory on array of hashes.
-        int positionsInStep = Math.min(addresses.size() + 1, (int) CACHE_SIZE.toBytes() / Integer.SIZE);
-        long[] positionToFullHashes = new long[positionsInStep];
-
-        for (int step = 0; step * positionsInStep <= addresses.size(); step++) {
-            int stepBeginPosition = step * positionsInStep;
-            int stepEndPosition = Math.min((step + 1) * positionsInStep, addresses.size());
-            int stepSize = stepEndPosition - stepBeginPosition;
-
-            // First extract all hashes from blocks to native array.
-            // Somehow having this as a separate loop is much faster compared
-            // to extracting hashes on the fly in the loop below.
-            for (int position = 0; position < stepSize; position++) {
-                int realPosition = position + stepBeginPosition;
-                long hash = readHashPosition(realPosition);
-                positionToFullHashes[position] = hash;
-                positionToHashes[realPosition] = (byte) hash;
-            }
-
-            // index pages
-            for (int position = 0; position < stepSize; position++) {
-                int realPosition = position + stepBeginPosition;
-                if (isPositionNull(realPosition)) {
-                    continue;
-                }
-
-                long hash = positionToFullHashes[position];
-                int pos = getHashPosition(hash, mask);
-
-                // look for an empty slot or a slot containing this key
-                while (key[pos] != -1) {
-                    int currentKey = key[pos];
-                    if (((byte) hash) == positionToHashes[currentKey] && positionEqualsPositionIgnoreNulls(currentKey, realPosition)) {
-                        // found a slot for this key
-                        // link the new key position to the current key position
-                        positionLinks[realPosition] = currentKey;
-
-                        // key[pos] updated outside of this loop
-                        break;
-                    }
-                    // increment position and mask to handler wrap around
-                    pos = (pos + 1) & mask;
-                }
-
-                key[pos] = realPosition;
-            }
-        }
-
-        size = sizeOf(addresses.elements()) + pagesHashStrategy.getSizeInBytes() +
-                sizeOf(key) + sizeOf(positionLinks) + sizeOf(positionToHashes);
     }
 
     @Override
     public final int getChannelCount()
     {
-        return channelCount;
+        return pagesHash.getChannelCount();
     }
 
     @Override
     public int getJoinPositionCount()
     {
-        return positionLinks.length;
+        return pagesHash.getPositionCount();
     }
 
     @Override
     public long getInMemorySizeInBytes()
     {
-        return size;
+        return pagesHash.getInMemorySizeInBytes();
     }
 
     @Override
     public long getJoinPosition(int position, Page hashChannelsPage, Page allChannelsPage)
     {
-        return getJoinPosition(position, hashChannelsPage, allChannelsPage, pagesHashStrategy.hashRow(position, hashChannelsPage));
+        int addressIndex = pagesHash.getAddressIndex(position, hashChannelsPage, allChannelsPage);
+        if (addressIndex == -1) {
+            return -1;
+        }
+        return getNextJoinPositionFrom(addressIndex, position, allChannelsPage);
     }
 
     @Override
-    public long getJoinPosition(int rightPosition, Page hashChannelsPage, Page allChannelsPage, long rawHash)
+    public long getJoinPosition(int position, Page hashChannelsPage, Page allChannelsPage, long rawHash)
     {
-        int pos = getHashPosition(rawHash, mask);
-
-        while (key[pos] != -1) {
-            if (positionEqualsCurrentRowIgnoreNulls(key[pos], (byte) rawHash, rightPosition, hashChannelsPage)) {
-                return getNextJoinPositionFrom(key[pos], rightPosition, allChannelsPage);
-            }
-            // increment position and mask to handler wrap around
-            pos = (pos + 1) & mask;
+        int addressIndex = pagesHash.getAddressIndex(position, hashChannelsPage, allChannelsPage, rawHash);
+        if (addressIndex == -1) {
+            return -1;
         }
-        return -1;
+        return getNextJoinPositionFrom(addressIndex, position, allChannelsPage);
     }
 
     @Override
     public final long getNextJoinPosition(long currentJoinPosition, int probePosition, Page allProbeChannelsPage)
     {
-        return getNextJoinPositionFrom(positionLinks[Ints.checkedCast(currentJoinPosition)], probePosition, allProbeChannelsPage);
+        int nextAddressIndex = pagesHash.getNextAddressIndex(Ints.checkedCast(currentJoinPosition));
+        return getNextJoinPositionFrom(nextAddressIndex, probePosition, allProbeChannelsPage);
     }
 
-    private long getNextJoinPositionFrom(int startJoinPosition, int probePosition, Page allProbeChannelsPage)
+    private int getNextJoinPositionFrom(int currentJoinPosition, int probePosition, Page allProbeChannelsPage)
     {
-        int currentJoinPosition = Ints.checkedCast(startJoinPosition);
         while (filterFunction != null && currentJoinPosition != -1 && !filterFunction.filter((currentJoinPosition), probePosition, allProbeChannelsPage)) {
-            currentJoinPosition = positionLinks[Ints.checkedCast(currentJoinPosition)];
+            currentJoinPosition = pagesHash.getNextAddressIndex(currentJoinPosition);
         }
         return currentJoinPosition;
     }
@@ -183,72 +102,6 @@ public final class JoinHash
     @Override
     public void appendTo(long position, PageBuilder pageBuilder, int outputChannelOffset)
     {
-        long pageAddress = addresses.getLong(Ints.checkedCast(position));
-        int blockIndex = decodeSliceIndex(pageAddress);
-        int blockPosition = decodePosition(pageAddress);
-
-        pagesHashStrategy.appendTo(blockIndex, blockPosition, pageBuilder, outputChannelOffset);
-    }
-
-    private boolean isPositionNull(int position)
-    {
-        long pageAddress = addresses.getLong(position);
-        int blockIndex = decodeSliceIndex(pageAddress);
-        int blockPosition = decodePosition(pageAddress);
-
-        return pagesHashStrategy.isPositionNull(blockIndex, blockPosition);
-    }
-
-    private long readHashPosition(int position)
-    {
-        long pageAddress = addresses.getLong(position);
-        int blockIndex = decodeSliceIndex(pageAddress);
-        int blockPosition = decodePosition(pageAddress);
-
-        return pagesHashStrategy.hashPosition(blockIndex, blockPosition);
-    }
-
-    private boolean positionEqualsCurrentRowIgnoreNulls(int leftPosition, byte rawHash, int rightPosition, Page rightPage)
-    {
-        if (positionToHashes[leftPosition] != rawHash) {
-            return false;
-        }
-
-        long pageAddress = addresses.getLong(leftPosition);
-        int blockIndex = decodeSliceIndex(pageAddress);
-        int blockPosition = decodePosition(pageAddress);
-
-        return pagesHashStrategy.positionEqualsRowIgnoreNulls(blockIndex, blockPosition, rightPosition, rightPage);
-    }
-
-    private boolean positionEqualsPositionIgnoreNulls(int leftPosition, int rightPosition)
-    {
-        long leftPageAddress = addresses.getLong(leftPosition);
-        int leftBlockIndex = decodeSliceIndex(leftPageAddress);
-        int leftBlockPosition = decodePosition(leftPageAddress);
-
-        long rightPageAddress = addresses.getLong(rightPosition);
-        int rightBlockIndex = decodeSliceIndex(rightPageAddress);
-        int rightBlockPosition = decodePosition(rightPageAddress);
-
-        return pagesHashStrategy.positionEqualsPositionIgnoreNulls(leftBlockIndex, leftBlockPosition, rightBlockIndex, rightBlockPosition);
-    }
-
-    private static int getHashPosition(long rawHash, long mask)
-    {
-        // Avalanches the bits of a long integer by applying the finalisation step of MurmurHash3.
-        //
-        // This function implements the finalisation step of Austin Appleby's <a href="http://sites.google.com/site/murmurhash/">MurmurHash3</a>.
-        // Its purpose is to avalanche the bits of the argument to within 0.25% bias. It is used, among other things, to scramble quickly (but deeply) the hash
-        // values returned by {@link Object#hashCode()}.
-        //
-
-        rawHash ^= rawHash >>> 33;
-        rawHash *= 0xff51afd7ed558ccdL;
-        rawHash ^= rawHash >>> 33;
-        rawHash *= 0xc4ceb9fe1a85ec53L;
-        rawHash ^= rawHash >>> 33;
-
-        return (int) (rawHash & mask);
+        pagesHash.appendTo(Ints.checkedCast(position), pageBuilder, outputChannelOffset);
     }
 }
